scripts/linkedin/requirements.txt
# Function dependencies, for example:
# package>=version
pandas
google-cloud-bigquery
datetime
requests
tqdm